# docs/swagger.yaml
definitions:
...
  dto.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  dto.TokenRefreshRequest:
    description: Запрос на обновление access-токена с использованием refresh-токена
    properties:
      token:
        description: |-
          Refresh-токен (обязательное поле)
          Required: true
          Example: eyJhbGciOiJIUzI1NiIsInR...
        type: string
    required:
    - token
    type: object
...
  dto.UserLoginRequest:
    description: Данные для авторизации пользователя
    properties:
      email:
        description: |-
          Email пользователя (обязательное поле)
          Required: true
          Example: user@example.com
        type: string
      password:
        description: |-
          Пароль (обязательное поле)
          Required: true
          Example: mysecurepassword
        type: string
    required:
    - email
    - password
    type: object
  dto.UserRegisterRequest:
    description: Данные, необходимые для регистрации пользователя
    properties:
      email:
        description: |-
          Email пользователя (обязательное поле)
          Required: true
          Example: user@example.com
        type: string
      password:
        description: |-
          Пароль пользователя (минимум 6 символов, обязательное поле)
          Required: true
          Example: mysecurepassword
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.UserResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      role:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Выполняет вход и возвращает токены
      parameters:
      - description: Данные для входа
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Токены доступа
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Ошибка авторизации
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Аутентификация пользователя
      tags:
      - Users
  /users/me:
    get:
      description: Возвращает информацию о пользователе по его `userID`
      produces:
      - application/json
      responses:
        "200":
          description: Данные о пользователе
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Пользователь не аутентифицирован
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Информация о текущем пользователе
      tags:
      - Users
...
  /users/refresh:
    post:
      consumes:
      - application/json
      description: Обновляет `access_token` и `refresh_token` по действующему `refresh_token`
      parameters:
      - description: Refresh токен
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/dto.TokenRefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Новые токены
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Ошибка авторизации
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление токенов
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе
      parameters:
      - description: Данные для регистрации
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Пользователь зарегистрирован'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - Users
swagger: "2.0"